Fix: potential app crash
The app should never crash and we must try our best to keep it running
even if catastrophic errors occur.  In this commit we try to keep the
app running when an invalid state has been encountered.  In the future
there should be code implementing an error reporting procedure.

Should we add this to our policy though?  We need to talk about it.  Is
it really beneficial to continue to run our app in an invalid state?
Could bad things happen?