Fix: on certain devices places page matrix messes up
On certain devices, the photos on places page have different heights
(and width as discovered later), varying from tens of pixels to
hundreds, and looks really funny.  The culprit is setting layout width
as `WRAP_CONTENT` when it should be 0 instead as we are using a weight
here.  When I coded this I didn't really get the difference but 0 should
stand for automatic.  Plus, the sizes are fully automatic.  Setting the
size (either width or height) to be the same as its content before the
sizes are calculated makes zero sense and is definitely not what we
want.