Fix: empty cards on main page and refactor API
Refactored API of the HorizontalRecyclerAdapter to use the setter
(for support of async loading) and fixed empty main page while I am
at it.  The next steps are to add the missing landmarks to the database
so we can remove the default image hack and refactor the database to be
its own thing, possibly put a global reference to database to the whole
application and then the HorizontalRecyclerAdapter would be able to find
the URLs itself instead of relying on the one passed in.  This could
reduce duplicate code and eliminate the need for the one method in the
Utils class.