Main page: use new name on tab icon
Keeping the shadow properly was hard...  This increases the bugged
non-hitbox area (that does not transmit event to the next layer) by a
tiny bit (the size of the shadow).  Note that the two images have
shadows that are not exactly the same so one of them might appear to be
cut off, and the image as a whole is asymmetric by one pixel.